# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: airmux Control Plane API
  description: 'Manage airmux organizations, workspaces, management keys, provider credentials, and data-plane synchronization.


    Organization and workspace targets are part of each URL. Successful responses wrap their result in

    `{"data": ...}`. Authenticate with a control-plane management key or a browser session as documented by

    each operation. Inference keys authenticate model requests at the data plane and are not accepted here.

    '
  version: 0.1.0
paths:
...
  /api/v1/service-accounts:
    post:
      tags:
      - Instance Users
      summary: Create Service Account
      description: 'Create a machine principal with an optional instance role.


        Required permission: `principals.manage`.'
      operationId: create_service_account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAccountIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      security:
      - ManagementKey: []
      - SessionCookie: []
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.manage
  /api/v1/service-accounts/{user_id}/management-keys:
    post:
      tags:
      - Instance Users
      summary: Create Instance Service Account Management Key
      description: 'Issue an instance key for an instance-managed service account.


        Required permission: `management-keys.issue`.'
      operationId: create_instance_service_account_management_key
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagementKeyIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ManagementKeyCreatedOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - management-keys.issue
  /api/v1/users/{user_id}:
    get:
      tags:
      - Instance Users
      summary: Get User
      description: 'Return one human user or service account and its organization memberships.


        Required permission: `principals.read`.'
      operationId: get_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.read
    delete:
      tags:
      - Instance Users
      summary: Delete User
      description: 'Delete a principal and its login identities, sessions, and control-plane management keys.


        Remove organization memberships, personal organizations, and workspaces containing inference

        keys created by this principal before deleting it.


        Required permission: `principals.manage`.'
      operationId: delete_user
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_DeletedOut_UUID__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.manage
  /api/v1/users/{user_id}/organizations:
    get:
      tags:
      - Instance Users
      summary: List User Organizations
      description: 'List one principal''s organization memberships.


        Required permission: `principals.read`.'
      operationId: list_user_organizations
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CursorToken'
          - type: 'null'
          title: Cursor
        description: Opaque continuation token from the previous page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEnvelope_MembershipOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.read
  /api/v1/users:
    get:
      tags:
      - Instance Users
      summary: List Users
      description: 'List human users and service accounts across the instance.


        Required permission: `principals.read`.'
      operationId: list_users
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: service_account
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Service Account
        description: 'Filter by principal type: true for service accounts and false for human users'
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CursorToken'
          - type: 'null'
          title: Cursor
        description: Opaque continuation token from the previous page
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: Maximum number of results to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEnvelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.read
  /api/v1/users/{user_id}/instance-role:
    put:
      tags:
      - Instance Users
      summary: Change Instance Role
      operationId: change_instance_role
      security:
      - ManagementKey: []
      - SessionCookie: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: User or service-account ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstanceRoleIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_UserOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Authentication failed
        '403':
          description: The credential does not have the required permission
      x-airmux-authority:
      - scope: instance_scope
        anyOf:
        - principals.manage
      description: 'Required permission: `principals.manage`.'
...
components:
  schemas:
...
    CursorToken:
      type: string
      maxLength: 512
      minLength: 1
      pattern: ^[A-Za-z0-9_-]+$
...
    DeletedOut_UUID_:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        deleted_at:
          type: string
          format: date-time
          title: Deleted At
      type: object
      required:
      - id
      - deleted_at
      title: DeletedOut[UUID]
...
    Envelope_DeletedOut_UUID__:
      properties:
        data:
          $ref: '#/components/schemas/DeletedOut_UUID_'
      type: object
      required:
      - data
      title: Envelope[DeletedOut[UUID]]
...
    Envelope_ManagementKeyCreatedOut_:
      properties:
        data:
          $ref: '#/components/schemas/ManagementKeyCreatedOut'
      type: object
      required:
      - data
      title: Envelope[ManagementKeyCreatedOut]
...
    Envelope_UserOut_:
      properties:
        data:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - data
      title: Envelope[UserOut]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InstanceRole:
      type: string
      enum:
      - owner
      - auditor
      - data_plane
      title: InstanceRole
    InstanceRoleIn:
      properties:
        instance_role:
          anyOf:
          - $ref: '#/components/schemas/InstanceRole'
          - type: 'null'
          description: Instance-wide role to assign, or null to remove instance-wide access
      additionalProperties: false
      type: object
      required:
      - instance_role
      title: InstanceRoleIn
...
    ManagementKeyCreatedOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        org_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Org Id
        workspace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workspace Id
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        prefix:
          type: string
          title: Prefix
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          title: Permissions
        label:
          type: string
          title: Label
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        scope:
          $ref: '#/components/schemas/Scope'
        status:
          type: string
          enum:
          - active
          - expired
          - revoked
          title: Status
        token:
          type: string
          title: Token
      type: object
      required:
      - id
      - user_id
      - org_id
      - workspace_id
      - parent_id
      - prefix
      - permissions
      - label
      - expires_at
      - revoked_at
      - created_at
      - updated_at
      - deleted_at
      - scope
      - status
      - token
      title: ManagementKeyCreatedOut
    ManagementKeyIn:
      properties:
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          minItems: 1
          title: Permissions
          description: Explicit maximum permissions carried by the key
        label:
          type: string
          maxLength: 80
          minLength: 1
          title: Label
          description: Where this key lives, such as ci, laptop, or data-plane
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
          description: Optional expiration timestamp with a timezone
      additionalProperties: false
      type: object
      required:
      - permissions
      - label
      title: ManagementKeyIn
...
    MembershipOut:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        org_id:
          type: string
          format: uuid
          title: Org Id
        role:
          $ref: '#/components/schemas/OrgRole'
        status:
          type: string
          const: member
          title: Status
      type: object
      required:
      - user_id
      - org_id
      - role
      - status
      title: MembershipOut
...
    OrgRole:
      type: string
      enum:
      - owner
      - admin
      - member
      - data_plane
      title: OrgRole
...
    PageEnvelope_MembershipOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/MembershipOut'
          type: array
          maxItems: 200
          title: Data
        page:
          $ref: '#/components/schemas/PageInfo'
      type: object
      required:
      - data
      - page
      title: PageEnvelope[MembershipOut]
...
    PageEnvelope_UserOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/UserOut'
          type: array
          maxItems: 200
          title: Data
        page:
          $ref: '#/components/schemas/PageInfo'
      type: object
      required:
      - data
      - page
      title: PageEnvelope[UserOut]
...
    PageInfo:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - next_cursor
      title: PageInfo
...
    Permission:
      type: string
      enum:
      - organizations.read
      - organizations.create
      - organizations.update
      - organizations.delete
      - principals.read
      - principals.manage
      - members.read
      - members.manage
      - workspaces.read
      - workspaces.create
      - workspaces.update
      - workspaces.delete
      - catalog.read
      - catalog.manage
      - provider-credentials.read
      - provider-credentials.manage
      - inference-keys.read
      - inference-keys.manage
      - policies.read
      - policies.manage
      - playground.execute
      - bundles.read
      - bundles.publish
      - usage.read
      - usage.ingest
      - data-planes.read
      - data-planes.heartbeat
      - audit.read
      - management-keys.read
      - management-keys.issue
      - management-keys.revoke
      title: Permission
...
    Scope:
      properties:
        level:
          $ref: '#/components/schemas/ScopeLevel'
        org_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Org Id
        workspace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workspace Id
      type: object
      required:
      - level
      title: Scope
    ScopeLevel:
      type: string
      enum:
      - instance
      - org
      - workspace
      title: ScopeLevel
...
    ServiceAccountIn:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
          description: Display name for the service account
        instance_role:
          anyOf:
          - $ref: '#/components/schemas/InstanceRole'
          - type: 'null'
          description: Optional instance-wide role for the service account
      additionalProperties: false
      type: object
      required:
      - name
      title: ServiceAccountIn
...
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        name:
          type: string
          title: Name
        instance_role:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Role
        service_account:
          type: boolean
          title: Service Account
        managing_org_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Managing Org Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        org_count:
          type: integer
          title: Org Count
      type: object
      required:
      - id
      - email
      - name
      - instance_role
      - service_account
      - managing_org_id
      - created_at
      - updated_at
      - deleted_at
      - org_count
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ManagementKey:
      type: http
      description: A control-plane management key using the `sk-cp-` prefix. Inference keys are not accepted by control-plane
        endpoints.
      scheme: bearer
    SessionCookie:
      type: apiKey
      in: cookie
      name: airmux_session
      description: Browser session cookie returned by login or signup. Browser requests must also send `X-Requested-With`.
...